index seems to work but...
pulling from the symtab needs reworking. Instead of copying and destroying
all the lists*, it would be better to make a new symbol pointing straight to
the list. This enables assignment to indeces. But destroying the parse tree
would need reworking. Perhaps only recursing in LibSym.destroySym() when
sym.tag is null. I tried that but it didn't work. But I didn't spend any time
looking for why. I think it's a good solution but I have to find what didn't
work. Also, Eval.index() is still in a "quick and dirty" state.

* and strings

 On branch n2t9
 Changes to be committed:
	modified:   Eval.jack
	modified:   Exec.jack
	modified:   LibSym.jack
	modified:   Parse.jack