bugs I thought were fixed, something's fishy
in Eval.formatExpression():

On no argument to a format expression, it would try to deallocate a null string.

On more than none but not enough arguments to format expression, it would
try to read 1 as apointer to a Symbol.  That was a problem with
Collection.nth().  Collection._nth() allows an overrun of 1 for internal use
but nth() shouldn't and it did.

In Parse.index(), I thought it handled recursive idences but it didn't.

In Exec.doRepeatIf, the call to Eval.condition() will cause side effects if
the argument points to a Collection.  It's a bug and I don't know how I want to
fix it.  A deep copy would work but that is very expensive.

Really, I thought I got (some of) these issues...

 On branch master
 Changes to be committed:
	modified:   LibSym.jack